refactor: Refactor directories (src)
