+++ jest/globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoDBName":"test","mongoUri":"mongodb://127.0.0.1:64164/"}
+{"mongoDBName":"test","mongoUri":"mongodb://127.0.0.1:63634/"}
