+++ package.json
@@ -55,6 +55,6 @@
     "prettier": "2.3.2",
     "tailwindcss": "2.2.14",
     "ts-jest": "27.0.5",
-    "typescript": "4.3.4"
+    "typescript": "4.2.3"
   }
 }
