+++ src/frontend/components/core/index.ts
