+++ src/frontend/hooks/use-is-initial-render.ts
