+++ src/pages/_app.tsx
@@ -7,8 +7,8 @@ import NProgress from 'nprogress';
 import { useEffect } from 'react';
 import { SWRConfig } from 'swr';
 
-import { CommonLayout } from '@src/components/layout';
-import { fetcher } from '@src/lib/fetcher';
+import { CommonLayout } from '@src/frontend/components/layout';
+import { fetcher } from '@src/frontend/lib/fetcher';
 
 import type { AppProps } from 'next/app';
