+++ src/pages/api/index.ts
@@ -1,4 +1,4 @@
-import { withErrorHandler } from '@utils/with-error-handler';
+import { withErrorHandler } from '@src/utils/with-error-handler';
 
 import type { NextApiRequest, NextApiResponse } from 'next';
