+++ src/pages/api/version.ts
