+++ src/utils/mongodb/encrypt.test.ts
@@ -1,7 +1,6 @@
 import { ObjectId } from 'bson';
 
-import { EncodeId } from 'types';
-
+import { EncodeId } from '@src/types';
 import { decodeId, encodeDocument, encodeDocuments, encodeId } from '@src/utils/mongodb/encrypt';
 
 describe('encrypt document', () => {
